Rewrote my proposal for the DSL
I tried to make a compact language that centers around the idea of running functions on devices with device:funtion syntax. And using variables to get and set state.

My focus was mostly to get all the functions required for the logic we have discussed, while keeping the syntax concise.

Most of the symbols can be replaced with words if a less compact DSL is preferred.